services/ingestion/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = "Multi-provider email ingestion service (Gmail + Outlook) for mail-manager"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.4.0",
    "asyncpg>=0.30.0",
    "httpx>=0.28.0",
    "redis[hiredis]>=5.2.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.160.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.14.0",
    "msal>=1.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ingestion"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
